fix: include POSITIONAL_ONLY params in _get_arg_keywords and improve docs
Address PR review recommendations:
- Include POSITIONAL_ONLY parameters in _get_arg_keywords() to fully match
  getfullargspec().args behavior (fixes potential telemetry arg mapping bug)
- Update get_func_parameters() docstring to note that annotations are strings
  on Python 3.14+
- Move get_func_parameters PEP 649 test to type_util_test.py for better
  organization
- Add test for positional-only parameter handling
